requirements.txt
# Requires python3.6+
# To install run "pip3 install -r requirements.txt" (or 'pip' instead of 'pip3' if 'pip' is for python3)

apturl==0.5.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
boto==2.49.0
boto3==1.9.47
botocore==1.12.47
Brlapi==0.6.6
bz2file==0.98
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
command-not-found==0.3
cryptography>=2.3
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
docutils==0.14
funcy==1.11
future==0.17.1
gensim==3.6.0
httplib2==0.9.2
idna==2.7
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
numexpr==2.6.8
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.8.0
protobuf==3.0.0
psycopg2==2.7.6.1
py==1.7.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyLDAvis==2.1.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pytest==4.0.0
python-apt==1.6.2
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.20.1
requests-unixsocket==0.1.5
s3transfer==0.1.13
scipy==1.1.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
smart-open==1.7.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
